Sauvegarde basculé dans le controller
